[BEAM-9618] Pull bundle descriptors for Go.
Support pulling bundle descriptors from the runner, as opposed to waiting for them to be pushed.

Cache bundle descriptors in the SDK so that multiple bundles with different instruction ids Ids in a given instance can share the same descriptors. This is valuable for Streaming jobs where registering the same bundle descriptor is common for efficiency.

Note, this doesn't add any garbage collection or clean up of older bundle ids, nor are teardown methods called for execution nodes, as before.