[BEAM-9733] Always let ImpulseSourceFunction emit a final watermark
The Flink Runner's ImpulseSourceFunction does not emit a final watermark, unless
the `--shutdownSourcesOnFinalWatermark` flag has been specified (the flag is
used in tests to shutdown the pipeline after reading all data). Most pipelines
will be long-running and thus do not specify the flag.

Not sending out the final watermark causes GroupByKey to hold back the data of
event time windows until the pipeline is shut down (the final watermark is
always emitted on pipeline shutdown which is why using the above flag works).