Use ref-counted connection to Spanner to prevent multiple connections.
In a multi-threaded runner, WriteToSpannerFn.setup() is called for
each instance on each thread, causing multiple session pools to be
created.

Use a ref-counted SpannerAccessor per database to ensure that there
is only one session pool per database per process.

Put the refcounting in SpannerAccessor so that it will be used
with all Spanner connections.